+++ Sources/RichTextKit/Sharing/StandardRichTextShareService.swift
@@ -54,7 +54,7 @@ public class StandardRichTextShareService: RichTextShareService {
             withName: fileName,
             extension: format.standardFileExtension,
             in: directory)
-        let data = try content.richTextData(with: format)
+        let data = try content.richTextData(for: format)
         try data.write(to: fileUrl)
         return fileUrl
     }