+++ Sources/RichTextKit/_Deprecated/RichTextDataWriter+Deprecated.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public extension RichTextDataWriter {
+
+    @available(*, deprecated, renamed: "richTextData(for:)")
+    func richTextData(with format: RichTextDataFormat) throws -> Data {
+        try richTextData(for: format)
+    }
+}