+++ RELEASE_NOTES.md
@@ -13,6 +13,10 @@ Until then, features may be deprecated in a minor version and removed in the nex
 * String extensions have been made public.
 * `RichTextDataFormat` has a new vendor-specific data format.
 
+### 🗑️ Deprecated
+
+* `RichTextDataWriter` `richTextData(with:)` is renamed to `richTextData(for:)`.  
+
 
 
 ## 0.1.2